refactor: move request and response types into separate header
