+++ include/neuron/define.h
@@ -86,4 +86,7 @@ typedef enum {
     NEU_TAG_CACHE_TYPE_NEVER,
 } neu_tag_cache_type_e;
 
+// forward declaration for neu_adapter_t
+typedef struct neu_adapter neu_adapter_t;
+
 #endif
